+++ test/svlog/items/assert_final.sv
@@ -1,7 +1,0 @@
-// RUN: moore %s -e foo
-// IGNORE  see issue #139
-
-module foo;
-    int x, y;
-    assert final (x == y);
-endmodule