+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Add support for `assert final` assertions.
+
 ### Fixed
 - Fix parsing of import declarations in module headers. ([#136](https://github.com/fabianschuiki/moore/issues/136))